business time

+++ marketplace/models/shoppingcart.model.js
@@ -16,7 +16,7 @@ var ShoppingCartModel = {
                 model.shoppingcart.products = products
                 model.shoppingcart.total = 0;
                 for (var i = model.shoppingcart.products.length - 1; i >= 0; i--) {
-                    model.shoppingcart.total = parseInt(model.shoppingcart.total) + parseInt(model.shoppingcart.products[i].price);
+                    model.shoppingcart.total = parseFloat(model.shoppingcart.total) + parseFloat(model.shoppingcart.products[i].price);
                 }
                 ;
